+++ src/test/java/org/woehlke/twitterwall/conf/TestConfig.java
@@ -1,0 +1,12 @@
+package org.woehlke.twitterwall.conf;
+
+import org.springframework.boot.context.properties.EnableConfigurationProperties;
+import org.springframework.context.annotation.Configuration;
+import org.woehlke.twitterwall.conf.properties.TestdataProperties;
+
+@Configuration
+@EnableConfigurationProperties({
+    TestdataProperties.class
+})
+public class TestConfig {
+}
