+++ src/test/java/org/woehlke/twitterwall/frontend/controller/ApplicationControllerTest.java
@@ -1,0 +1,81 @@
+package org.woehlke.twitterwall.frontend.controller;
+
+
+import org.junit.Assert;
+import org.junit.Ignore;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.annotation.Commit;
+import org.springframework.test.context.junit4.SpringRunner;
+import org.springframework.test.web.servlet.MockMvc;
+import org.springframework.test.web.servlet.MvcResult;
+import org.woehlke.twitterwall.Application;
+import org.woehlke.twitterwall.PrepareDataTest;
+
+import static org.assertj.core.api.Java6Assertions.assertThat;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.model;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.view;
+
+@RunWith(SpringRunner.class)
+@SpringBootTest(classes={Application.class},webEnvironment=SpringBootTest.WebEnvironment.RANDOM_PORT)
+@AutoConfigureMockMvc
+public class ApplicationControllerTest {
+
+    private static final Logger log = LoggerFactory.getLogger(ApplicationControllerTest.class);
+
+    @Autowired
+    private MockMvc mockMvc;
+
+    @Autowired
+    private ApplicationController controller;
+
+    @Autowired
+    private PrepareDataTest prepareDataTest;
+
+    @Commit
+    @Test
+    public void controllerIsPresentTest(){
+        log.info("controllerIsPresentTest");
+        assertThat(controller).isNotNull();
+    }
+
+    @Commit
+    @Test
+    public void setupTestData() throws Exception {
+        String msg = "setupTestData: ";
+        prepareDataTest.getTestDataTweets(msg);
+        prepareDataTest.getTestDataUser(msg);
+        Assert.assertTrue(true);
+    }
+
+    //TODO: #180 https://github.com/phasenraum2010/twitterwall2/issues/180
+    @Ignore
+    @Commit
+    @Test
+    public void domainCountTest() throws Exception {
+        String msg ="domainCountTest: ";
+
+        MvcResult result = this.mockMvc.perform(get("/application/domain/count"))
+                .andExpect(status().isOk())
+                .andExpect(view().name( "application/domain/count"))
+                .andExpect(model().attributeExists("countedEntities"))
+                .andExpect(model().attributeExists("page"))
+                .andReturn();
+
+        String content = result.getResponse().getContentAsString();
+
+        log.info(msg+"#######################################");
+        log.info(msg+"#######################################");
+        log.info(msg+content);
+        log.info(msg+"#######################################");
+        log.info(msg+"#######################################");
+        Assert.assertTrue(true);
+    }
+}
