+++ src/test/java/org/woehlke/twitterwall/frontend/controller/ImprintControllerTest.java
@@ -1,6 +1,7 @@
 package org.woehlke.twitterwall.frontend.controller;
 
 import org.junit.Assert;
+import org.junit.Ignore;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.slf4j.Logger;
@@ -15,7 +16,7 @@ import org.springframework.test.web.servlet.MvcResult;
 import org.woehlke.twitterwall.Application;
 import org.woehlke.twitterwall.conf.properties.TwitterProperties;
 import org.woehlke.twitterwall.conf.properties.FrontendProperties;
-import org.woehlke.twitterwall.test.UserServiceTest;
+import org.woehlke.twitterwall.test.UserServiceTestHelper;
 
 import static org.assertj.core.api.Java6Assertions.assertThat;
 import static org.hamcrest.CoreMatchers.containsString;
@@ -42,7 +43,7 @@ public class ImprintControllerTest {
     private MockMvc mockMvc;
 
     @Autowired
-    private UserServiceTest userServiceTest;
+    private UserServiceTestHelper userServiceTestHelper;
 
     @Autowired
     private TwitterProperties twitterProperties;
@@ -59,10 +60,10 @@ public class ImprintControllerTest {
 
     @Commit
     @Test
-    public void fetchTweetsFromTwitterSearchTest() {
+    public void fetchTweetsFromTwitterSearchTest() throws Exception  {
         log.info("------------------------------------");
         log.info("fetchTweetsFromTwitterSearchTest: START  userServiceTest.createUser("+ frontendProperties.getImprintScreenName()+")");
-        userServiceTest.createUser(frontendProperties.getImprintScreenName());
+        userServiceTestHelper.createUser(frontendProperties.getImprintScreenName());
         log.info("fetchTweetsFromTwitterSearchTest: DONE  userServiceTest.createUser("+ frontendProperties.getImprintScreenName()+")");
         log.info("------------------------------------");
         Assert.assertTrue(true);
@@ -76,6 +77,8 @@ public class ImprintControllerTest {
         Assert.assertTrue(true);
     }
 
+    //TODO: #182 https://github.com/phasenraum2010/twitterwall2/issues/182
+    @Ignore
     @Commit
     @Test
     public void imprintTest2() throws Exception {
