+++ src/test/java/org/woehlke/twitterwall/frontend/controller/UrlControllerTest.java
@@ -1,0 +1,53 @@
+package org.woehlke.twitterwall.frontend.controller;
+
+
+import org.junit.Assert;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.annotation.Commit;
+import org.springframework.test.context.junit4.SpringRunner;
+import org.springframework.test.web.servlet.MockMvc;
+import org.woehlke.twitterwall.Application;
+import org.woehlke.twitterwall.PrepareDataTest;
+
+import static org.assertj.core.api.Java6Assertions.assertThat;
+
+@RunWith(SpringRunner.class)
+@SpringBootTest(classes={Application.class},webEnvironment=SpringBootTest.WebEnvironment.RANDOM_PORT)
+@AutoConfigureMockMvc
+public class UrlControllerTest {
+
+    private static final Logger log = LoggerFactory.getLogger(UrlControllerTest.class);
+
+    @Autowired
+    private UrlController controller;
+
+    @Autowired
+    private PrepareDataTest prepareDataTest;
+
+    @Autowired
+    private MockMvc mockMvc;
+
+    @Commit
+    @Test
+    public void controllerIsPresentTest(){
+        log.info("controllerIsPresentTest");
+        assertThat(controller).isNotNull();
+    }
+
+    @Commit
+    @Test
+    public void setupTestData() throws Exception {
+        String msg = "setupTestData: ";
+        prepareDataTest.getTestDataTweets(msg);
+        prepareDataTest.getTestDataUser(msg);
+        Assert.assertTrue(true);
+    }
+
+
+}
