+++ src/test/java/org/woehlke/twitterwall/oodm/entities/DomainObjectMinimalTest.java
@@ -1,0 +1,8 @@
+package org.woehlke.twitterwall.oodm.entities;
+
+public interface DomainObjectMinimalTest {
+
+    void getUniqueIdTest() throws Exception;
+
+    void isValidTest() throws Exception;
+}
