+++ src/test/java/org/woehlke/twitterwall/oodm/entities/TweetTest.java
@@ -1,0 +1,23 @@
+package org.woehlke.twitterwall.oodm.entities;
+
+import org.junit.Assert;
+import org.junit.Test;
+
+public class TweetTest implements DomainObjectMinimalTest  {
+
+    //TODO: #197 https://github.com/phasenraum2010/twitterwall2/issues/197
+    @Test
+    @Override
+    public void getUniqueIdTest() throws Exception {
+        String msg = "getUniqueIdTest: ";
+        Assert.assertTrue(msg,true);
+    }
+
+    //TODO: #197 https://github.com/phasenraum2010/twitterwall2/issues/197
+    @Test
+    @Override
+    public void isValidTest() throws Exception {
+        String msg = "isValidTest: ";
+        Assert.assertTrue(msg,true);
+    }
+}
