+++ src/test/java/org/woehlke/twitterwall/scheduled/mq/endoint/AsyncStartTaskTest.java
@@ -1,0 +1,18 @@
+package org.woehlke.twitterwall.scheduled.mq.endoint;
+
+
+/**
+ * @see org.woehlke.twitterwall.scheduled.mq.endoint.AsyncStartTask
+ */
+public interface AsyncStartTaskTest {
+
+    void updateTweetsTest() throws Exception;
+
+    void updateUserProfilesTest() throws Exception;
+
+    void updateUserProfilesFromMentionsTest() throws Exception;
+
+    void fetchTweetsFromTwitterSearchTest() throws Exception;
+
+    void fetchUsersFromDefinedUserListTest() throws Exception;
+}
