+++ src/test/java/org/woehlke/twitterwall/scheduled/service/backend/TwitterUrlServiceTestDataTest.java
@@ -18,7 +18,7 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
-import static org.woehlke.twitterwall.test.ScheduledTasksFacadeTest.PROFILE_URLS;
+import static org.woehlke.twitterwall.test.TestDataHelper.PROFILE_URLS;
 
 /**
  * Created by tw on 21.06.17.
