--- src/test/java/org/woehlke/twitterwall/test/ScheduledTasksFacadeTest.java
@@ -1,68 +1,0 @@
-package org.woehlke.twitterwall.test;
-
-/**
- * Created by tw on 25.06.17.
- */
-public interface ScheduledTasksFacadeTest {
-    
-    String PROFILE_URLS[] = {
-            "http://t.co/hIxL9WeoP1",
-            "http://t.co/FGRp2dcyII",
-            "https://t.co/t72OfGzzfd",
-            "http://t.co/gnmbZ3HLPC",
-            "http://t.co/5uLiz5mReg",
-            "https://t.co/Lp00eOlYYK",
-            "https://t.co/4g5pdzrWkM",
-            "https://t.co/feV9khihK3",
-            "https://t.co/Lq9HDdbM1S",
-            "http://t.co/ZK8jQ2KEJD",
-            "https://t.co/uzzouytYxh",
-            "http://t.co/7ccMD0G00p",
-            "http://t.co/mNdNHD0dLO",
-            "https://t.co/lQlse7u93G",
-            "https://t.co/f2vtNkiwTJ",
-            "http://t.co/sgSySSyjol",
-            "https://t.co/Fb3UEeVWWk",
-            "https://t.co/n9LlZXFjTc",
-            "http://t.co/a7oWdQhTin",
-            "http://t.co/8EuxwYs0va",
-            "http://t.co/75Hf3JV6JU",
-            "https://t.co/qIDUMVN3WY",
-            "https://t.co/kM9PIn3BgS",
-            "https://t.co/46Dz7qCB8L",
-            "http://t.co/RzkVNLQLx6",
-            "http://t.co/nWxOAD6WNC",
-            "http://t.co/9YEDGUjbqK",
-            "https://t.co/lQlse7u93G",
-            "https://t.co/4tDiIJ74eR",
-            "https://t.co/xnOcVzUTkm",
-            "https://t.co/Ycl7ZCbRsy",
-            "https://t.co/N7Rq7beH8n",
-            "https://t.co/okIuRGVRi1",
-            "https://t.co/OP7oBmBA3T",
-            "http://t.co/N0wDJifTl8",
-            "https://t.co/NWnpbiSrFg",
-            "http://t.co/yspEKRBcSw",
-            "https://t.co/MuVYy9ahOi",
-            "https://t.co/HQpP0PXtCY",
-            "http://t.co/9P2qeHaGH7",
-            "http://t.co/tjFrOTqr1l",
-            "https://t.co/vFHaMjhQxK",
-            "http://t.co/TC1JsRY0sY",
-            "https://t.co/DWliJ3cEUT",
-            "https://t.co/Yfcyk3aqh1",
-            "https://t.co/5s5ggTRpIX",
-            "https://t.co/z0lOKQCs0y",
-            "https://t.co/gnmbZ3qaY4",
-            "http://t.co/9Ct4HFvTSM",
-            "http://t.co/gnmbZ3qaY4",
-            "http://t.co/CudjEZGLVT",
-            "http://t.co/imtLrCUAko",
-            "https://t.co/RvKQvMMpzF",
-            "https://t.co/7haKo3blhj",
-            "http://t.co/MALJQIyYrt",
-            "https://t.co/4HBxydWdIE",
-            "https://t.co/hMMJBGFFKR",
-            "https://t.co/cbsEBiSW2E",
-    };
-}
