--- src/test/java/org/woehlke/twitterwall/test/UrlServiceTest.java
@@ -1,14 +1,0 @@
-package org.woehlke.twitterwall.test;
-
-import org.woehlke.twitterwall.oodm.entities.Task;
-import org.woehlke.twitterwall.oodm.entities.Url;
-
-import java.util.List;
-
-/**
- * Created by tw on 01.07.17.
- */
-public interface UrlServiceTest {
-
-    List<Url> getTestData(Task task);
-}
