+++ src/test/java/org/woehlke/twitterwall/test/UserServiceTestHelper.java
@@ -1,0 +1,16 @@
+package org.woehlke.twitterwall.test;
+
+import org.woehlke.twitterwall.oodm.entities.User;
+
+/**
+ * Created by tw on 01.07.17.
+ */
+@Deprecated
+public interface UserServiceTestHelper {
+
+    @Deprecated
+    User createImprintUser();
+
+    @Deprecated
+    User createUser(String username);
+}
